moved cred2 components into own module, removed TokenPruneService
