+++ broker-auth-multiuser-otp/src/main/java/org/aktin/broker/auth/otp/http/AuthEndpoint.java
@@ -1,4 +1,4 @@
-package org.aktin.broker.auth.cred2.http;
+package org.aktin.broker.auth.otp.http;
 
 import java.util.concurrent.TimeUnit;
 import java.util.logging.Logger;
@@ -13,9 +13,9 @@ import javax.ws.rs.Produces;
 import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
-import org.aktin.broker.auth.cred2.auth.Token;
-import org.aktin.broker.auth.cred2.auth.TokenManager;
-import org.aktin.broker.auth.cred2.service.UserAuthService;
+import org.aktin.broker.auth.otp.auth.Token;
+import org.aktin.broker.auth.otp.auth.TokenManager;
+import org.aktin.broker.auth.otp.service.UserAuthService;
 import org.aktin.broker.rest.Authenticated;
 import org.aktin.broker.rest.RequireAdmin;
 import org.aktin.broker.server.auth.HttpBearerAuthentication;
