+++ broker-auth-multiuser-otp/src/main/java/org/aktin/broker/auth/otp/http/Credentials.java
@@ -1,4 +1,4 @@
-package org.aktin.broker.auth.cred2.http;
+package org.aktin.broker.auth.otp.http;
 
 import java.io.Serializable;
 import javax.xml.bind.annotation.XmlAccessType;
@@ -7,11 +7,10 @@ import javax.xml.bind.annotation.XmlRootElement;
 
 @XmlRootElement
 @XmlAccessorType(XmlAccessType.FIELD)
-public class Status implements Serializable {
+public class Credentials implements Serializable {
 
-  private static final long serialVersionUID = 2L;
+  private static final long serialVersionUID = 1L;
 
   public String username;
-  public long issued;
-  public long expiresAt;
+  public String password;
 }
