+++ broker-auth-multiuser-otp/src/main/java/org/aktin/broker/auth/otp/http/Status.java
@@ -1,4 +1,4 @@
-package org.aktin.broker.auth.cred2.http;
+package org.aktin.broker.auth.otp.http;
 
 import java.io.Serializable;
 import javax.xml.bind.annotation.XmlAccessType;
@@ -7,10 +7,11 @@ import javax.xml.bind.annotation.XmlRootElement;
 
 @XmlRootElement
 @XmlAccessorType(XmlAccessType.FIELD)
-public class Credentials implements Serializable {
+public class Status implements Serializable {
 
-  private static final long serialVersionUID = 1L;
+  private static final long serialVersionUID = 2L;
 
   public String username;
-  public String password;
+  public long issued;
+  public long expiresAt;
 }
