+++ broker-auth-multiuser-otp/src/main/java/org/aktin/broker/auth/otp/repository/JdbcUserRepository.java
@@ -1,4 +1,4 @@
-package org.aktin.broker.auth.cred2.repository;
+package org.aktin.broker.auth.otp.repository;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
