+++ broker-auth-multiuser-otp/src/main/java/org/aktin/broker/auth/otp/repository/User.java
@@ -1,6 +1,5 @@
-package org.aktin.broker.auth.cred2.repository;
+package org.aktin.broker.auth.otp.repository;
 
-import java.sql.Timestamp;
 import java.util.Objects;
 
 /**
