+++ broker-auth-multiuser-otp/src/main/java/org/aktin/broker/auth/otp/service/JdbcUserService.java
@@ -1,4 +1,4 @@
-package org.aktin.broker.auth.cred2.service;
+package org.aktin.broker.auth.otp.service;
 
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -11,9 +11,9 @@ import java.util.logging.Logger;
 import javax.inject.Inject;
 import javax.inject.Singleton;
 import javax.sql.DataSource;
-import org.aktin.broker.auth.cred2.repository.User;
-import org.aktin.broker.auth.cred2.repository.UserRepository;
-import org.aktin.broker.auth.cred2.utils.PasswordHasher;
+import org.aktin.broker.auth.otp.repository.User;
+import org.aktin.broker.auth.otp.repository.UserRepository;
+import org.aktin.broker.auth.otp.utils.PasswordHasher;
 
 @Singleton
 public class JdbcUserService implements UserService {
